pyproject.toml
[project]
name = "persistedstate"
version = "25.1"
description = "Ultra fast persistent store for small states in mapped YAML file"
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Utilities",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = ["pyyaml"]

[dependency-groups]
dev = [
    "black",
    "mypy",
    "pylint",
    "scripthelper",
    "tox",
    "tox-uv>=1.25",
    "types-PyYAML",
    "pytest"
]
perftest = [
    "diskcache",
    "lmdbm",
    "sqlitedict",
]

[[project.authors]]
name = "Máté Farkas"
email = "fm@farkas-mate.hu"

[project.urls]
Homepage = "https://github.com/presidento/persistedstate"
Changelog = "https://github.com/presidento/persistedstate/blob/main/CHANGELOG.md"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[tool.tox]
env_list = ["py314", "py313", "py312", "py311", "py310"]

[tool.tox.env_run_base]
commands = [["pytest"]]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
no_site_packages = true

[[tool.mypy.overrides]]
module = "persistedstate"
disable_error_code = ["override", "attr-defined"]

[tool.pylint.main]
disable = [
    "attribute-defined-outside-init",
    "consider-using-with",
    "format",
    "invalid-name",
    "logging-fstring-interpolation",
    "logging-not-lazy",
    "missing-docstring",
    "no-member",
    "protected-access",
]
